Update markup inside sidebar and main element
This changes <aside> to a regular <div>, and specifies that the <nav> element is for main navigation. It also swaps the <h2> header elements used in the navigation in favor of <div> tags, and changes the <cite> to <div>. Finally the "main" div is using <main> so it can get picked up as a proper landmark.

Fixes https://www.pivotaltracker.com/story/show/169826719