Improve store name display on mobile and fix sidebar overlapping
The line height and font size in the sidebar were off on mobile, this change makes sure they're set to the correct values. It also fixes a bug with the sidebar content overflowing and appearing underneath the normal page content, when there's a large number of links in the sidebar.